# YJIT: Reorder branches for Fixnum opt_case_dispatch